% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{assign_policy_label}
\alias{assign_policy_label}
\title{Assigns policy labels to data for use in create_policies_from_data}
\usage{
assign_policy_label(data, ...)
}
\description{
Assigns policy labels to data for use in create_policies_from_data
}
