% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{compute_loss}
\alias{compute_loss}
\title{Loss given a specific pooling stucture (sigma)}
\usage{
compute_loss(
  data,
  value,
  M,
  sigma,
  policy_list,
  reg = 1,
  normalize = 0,
  lattice_edges = NA,
  R
)
}
\arguments{
\item{data}{User supplied dataframe that contains values}

\item{value}{Label of column containing value observed for each observation}

\item{M}{Dataframe of policy means}

\item{sigma}{Partition matrix that gives pooling structure}

\item{policy_list}{List of policies, which if a list of vectors of all policies implied
by the data.}

\item{reg}{Regularization parameter that penalizes partitions with more pools}

\item{normalize}{Whether or not to normalize loss (unsure?)}

\item{lattice_edges}{Edges of pooling structure}

\item{R}{A list (or integer) of the number of levels in each arm.}

\item{i}{Row to split at in sigma}

\item{j}{Column to split from in sigma}
}
\value{
Loss given pool for the data
}
\description{
Loss given a specific pooling stucture (sigma)
}
