% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{connected_components}
\alias{connected_components}
\title{Returns the pooling structure given a sigma, in the form of a list that gives
the pool number for each factorial combination (policy id)}
\usage{
connected_components(n, edges, policy_list)
}
\arguments{
\item{n}{The initial number of parent nodes, usually the total number of factorial combinations}

\item{edges}{The edgelist of the Hasse diagram given by lattice_edges or prune_edges}
}
\value{
A collections:dict() that gives the pool for each factorial combination (policy id). The key
for each element of the list is the index of its occurrence in policy_list (given by create policies),
and its value is the pool it is assigned to.
}
\description{
Returns the pooling structure given a sigma, in the form of a list that gives
the pool number for each factorial combination (policy id)
}
