% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{create_policies_from_data}
\alias{create_policies_from_data}
\title{Creates all factor combinations in a factorial design (as in) create_policies}
\usage{
create_policies_from_data(data, ...)
}
\description{
Creates all factor combinations in a factorial design (as in) create_policies
}
