% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_feasible_combinations}
\alias{find_feasible_combinations}
\title{Finds feasible combinations of}
\usage{
find_feasible_combinations(rashomon_profiles, theta, H, sorted = FALSE)
}
\arguments{
\item{rashomon_profiles}{A list of RashomonSet objects, where each entry of the list
contains a RashomonSet corresponding to a different profile}

\item{theta}{Threshold value to be present in the RashomonSet.}

\item{H}{The maximum number of pools that can be present across all profiles.}

\item{sorted}{Whether or not the RashomonSet objects have been sorted by the internal $sort method.
Defaults to FALSE}
}
\value{
A list of feasible combinations of poolings such that the sum of the losses across
profiles is less than theta and has less than H total pools. The ith entry in
each list corresponds to the ith RashomonSet profile, and the value of
the ith entry represents the model in the ith RashomonSet profile.
}
\description{
Finds feasible combinations of
}
