% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_feasible_sum_subsets}
\alias{find_feasible_sum_subsets}
\title{Find all combinations of indices from the list of arrays S such that sum <= theta}
\usage{
find_feasible_sum_subsets(S, theta)
}
\arguments{
\item{S}{List of lists of numbers}

\item{theta}{Threshold}
}
\value{
List of combinations of indices, one from each array in S
such that their sum is less than or equal to theta
}
\description{
Find all combinations of indices from the list of arrays S such that sum <= theta
}
