% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_profile_lower_bound}
\alias{find_profile_lower_bound}
\title{Finds lower bound for a given profile}
\usage{
find_profile_lower_bound(data, value)
}
\arguments{
\item{data}{Data prefiltered to be from a given profile}

\item{value}{The column name of the y values in data}
}
\value{
Bound on how low a model's loss can be in a given profile.
}
\description{
Finds lower bound for a given profile
}
