% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RashomonSet.R
\name{insert_model}
\alias{insert_model}
\title{Inserts new model and it's corresponding loss into a RashomonSet Object}
\usage{
insert_model(new_sigma, loss, num_pools, profile)
}
\arguments{
\item{new_sigma}{A M X R-2 partition matrix}

\item{loss}{The loss of the model when using this pooling structure}

\item{num_pools}{Number of pools present in new_sigma}

\item{insert_profile}{Profile for the given sigma matrix.}
}
\description{
Inserts new model and it's corresponding loss into a RashomonSet Object
}
