% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{lattice_edges}
\alias{lattice_edges}
\title{Calculates edgelist for Hasse diagram given sigma and a list of factor combinations (policies).}
\usage{
lattice_edges(sigma, policy_list)
}
\arguments{
\item{sigma}{M X R-2 matrix that gives where cuts occur (an entry if 1 is no cut, and 0 if there is)}

\item{policy_list}{A list of all factorial combinations given by create_policies}
}
\value{
Edge list of Hasse diagram, specifying which policies are connected in the same pool
}
\description{
Calculates edgelist for Hasse diagram given sigma and a list of factor combinations (policies).
}
