% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{merge_components}
\alias{merge_components}
\title{helper function for use in union_find data structure}
\usage{
merge_components(parent, x)
}
\description{
helper function for use in union_find data structure
}
