% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{num_pools_change_r}
\alias{num_pools_change_r}
\title{Finds the number of pools in a sigma with differing numbers of levels for each arm.}
\usage{
num_pools_change_r(sigma, R)
}
\arguments{
\item{sigma}{Partition matrix for a given pooling structure}

\item{R}{Vector of the number of levels for each arm}
}
\value{
The number of pools given by this pooling structure
}
\description{
Finds the number of pools in a sigma with differing numbers of levels for each arm.
}
