% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{num_pools_fixed_r}
\alias{num_pools_fixed_r}
\title{Finds the number of pools in a sigma with the same number of levels for each arm.}
\usage{
num_pools_fixed_r(sigma, R)
}
\arguments{
\item{sigma}{Partition matrix for a given pooling structure}

\item{R}{The number of total numbers of levels for each arm (in this function,
we are assuming R is equal for all arms)}
}
\value{
The number of pools given by this pooling structure
}
\description{
Finds the number of pools in a sigma with the same number of levels for each arm.
}
