% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{partition_sigma}
\alias{partition_sigma}
\title{Maximally cuts in row i for use in the lookahead loss, up to column j.}
\usage{
partition_sigma(i, j, sigma)
}
\arguments{
\item{i}{Row of the sigma matrix to cut at}

\item{j}{Column to cut up to}

\item{sigma}{Partition matrix for a given pooling structure}
}
\value{
A new partition matrix cut from row i to column j
}
\description{
Maximally cuts in row i for use in the lookahead loss, up to column j.
}
