% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pol_in_prof}
\alias{pol_in_prof}
\title{Finds the policies present in a given profile.}
\usage{
pol_in_prof(policies, profile, inactive = 0)
}
\arguments{
\item{policies}{A list of policies, typically given by generate_policies_from_data}

\item{profile}{The profile we wish to see which policies are active form}

\item{inactive}{The level in policies that denotes an arm is inactive.}
}
\value{
A list of policies present in a given profile
}
\description{
Finds the policies present in a given profile.
}
