% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{pool_means}
\alias{pool_means}
\title{Finds pool means given data, value, and pool label
TODO: Change all function calls from other packages to be explicit, and
change if necessary for speedups.}
\usage{
pool_means(data, value, pool)
}
\arguments{
\item{data}{User data appended with policy labels}

\item{value}{Name of value column observed for each observation, supplied by the user.}

\item{pool}{Name of pool column}
}
\value{
A collections::dict() object, where the key is an integer i corresponding
to the pool id and its value is the mean of that pool.
}
\description{
Finds pool means given data, value, and pool label
TODO: Change all function calls from other packages to be explicit, and
change if necessary for speedups.
}
