% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{powerset}
\alias{powerset}
\title{Finds powerset of a given list (all the possible subsets of a list)}
\usage{
powerset(arr)
}
\arguments{
\item{arr}{A list}
}
\value{
A list of lists, where element list[\link{k}] is a matrix of all of
the subsets of size k, such that list[\link{k}][,i] is the ith subset of size k.
}
\description{
Finds powerset of a given list (all the possible subsets of a list)
}
