% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RashomonSet.R
\docType{class}
\name{RashomonSet-class}
\alias{RashomonSet-class}
\alias{RashomonSet}
\title{Rashomon Set}
\description{
This class stores members of the RashomonSet. This can be used to both form the RashomonSet for a given
profile and a single member of the RashomonSet across all profiles.
}
\section{Fields}{

\describe{
\item{\code{models}}{The M X R-2 partition matrices that give the pooling structure for
their profile}

\item{\code{losses}}{The losses for each of the models when evaluated on their profile}

\item{\code{pools}}{The number of pools in the each of the models when evaluated on their profile}

\item{\code{profiles}}{The profiles for each of the models.}
}}


