% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RashomonSet.R
\name{sort.RashomonSet}
\alias{sort.RashomonSet}
\title{Sorts this rashomon set according to the model with the smallest loss.}
\usage{
\method{sort}{RashomonSet}(obj)
}
\description{
Sorts this rashomon set according to the model with the smallest loss.
}
