% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RashomonSet.R
\name{sort}
\alias{sort}
\title{Sorts this rashomon set according to the model with the smallest loss.}
\usage{
sort()
}
\description{
Sorts this rashomon set according to the model with the smallest loss.
}
