% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_prof}
\alias{subset_prof}
\title{Finds subset of a dataset corresponding to a given profile}
\usage{
subset_prof(data, policy_list, profile, inactive = 0)
}
\arguments{
\item{data}{A dataframe output from assign_policy_label}

\item{policy_list}{A list of policies given by generate_policies_from_data}

\item{profile}{The profile we wish to see which policies are active form}

\item{inactive}{The level in policies that denotes an arm is inactive.}
}
\value{
The subset of a data that corresponds to the given profile
}
\description{
Finds subset of a dataset corresponding to a given profile
}
