% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{sum_product_k}
\alias{sum_product_k}
\title{Finds sum of all the products of size k for each k < n, where n is the size
of the list. Useful for counting the number of pools.}
\usage{
sum_product_k(arr)
}
\arguments{
\item{arr}{A list}
}
\value{
A list of the sums of the products of size k, where the (k+1)th element
denotes the kth sum. The 1st element is 1, for use in other functions for easy looping.
}
\description{
Finds sum of all the products of size k for each k < n, where n is the size
of the list. Useful for counting the number of pools.
}
